+++ crates/client/rpc/src/methods/read/get_transaction_receipt.rs
@@ -28,7 +28,7 @@ use crate::Starknet;
 ///
 /// The function may return a `TXN_HASH_NOT_FOUND` error if the specified transaction hash is
 /// not found.
-pub async fn get_transaction_receipt(
+pub fn get_transaction_receipt(
     starknet: &Starknet,
     transaction_hash: Felt,
 ) -> StarknetRpcResult<TransactionReceiptWithBlockInfo> {
@@ -64,4 +64,52 @@ pub async fn get_transaction_receipt(
     };
 
     Ok(TransactionReceiptWithBlockInfo { receipt, block })
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::test_utils::{sample_chain_for_block_getters, SampleChainForBlockGetters};
+    use rstest::rstest;
+    use starknet_core::types::ReceiptBlock;
+
+    #[rstest]
+    fn test_get_transaction_receipt(sample_chain_for_block_getters: (SampleChainForBlockGetters, Starknet)) {
+        let (SampleChainForBlockGetters { block_hashes, tx_hashes, expected_receipts, .. }, rpc) =
+            sample_chain_for_block_getters;
+
+        // Block 0
+        let res = TransactionReceiptWithBlockInfo {
+            receipt: expected_receipts[0].clone(),
+            block: ReceiptBlock::Block { block_hash: block_hashes[0], block_number: 0 },
+        };
+        assert_eq!(get_transaction_receipt(&rpc, tx_hashes[0]).unwrap(), res);
+
+        // Block 1
+
+        // Block 2
+        let res = TransactionReceiptWithBlockInfo {
+            receipt: expected_receipts[1].clone(),
+            block: ReceiptBlock::Block { block_hash: block_hashes[2], block_number: 2 },
+        };
+        assert_eq!(get_transaction_receipt(&rpc, tx_hashes[1]).unwrap(), res);
+        let res = TransactionReceiptWithBlockInfo {
+            receipt: expected_receipts[2].clone(),
+            block: ReceiptBlock::Block { block_hash: block_hashes[2], block_number: 2 },
+        };
+        assert_eq!(get_transaction_receipt(&rpc, tx_hashes[2]).unwrap(), res);
+
+        // Pending
+        let res =
+            TransactionReceiptWithBlockInfo { receipt: expected_receipts[3].clone(), block: ReceiptBlock::Pending };
+        assert_eq!(get_transaction_receipt(&rpc, tx_hashes[3]).unwrap(), res);
+    }
+
+    #[rstest]
+    fn test_get_transaction_receipt_not_found(sample_chain_for_block_getters: (SampleChainForBlockGetters, Starknet)) {
+        let (SampleChainForBlockGetters { .. }, rpc) = sample_chain_for_block_getters;
+
+        let does_not_exist = Felt::from_hex_unchecked("0x7128638126378");
+        assert_eq!(get_transaction_receipt(&rpc, does_not_exist), Err(StarknetRpcApiError::TxnHashNotFound));
+    }
 }