+++ crates/client/rpc/src/methods/trace/trace_transaction.rs
@@ -7,8 +7,8 @@ use dc_exec::execution_result_to_tx_trace;
 use dc_exec::ExecutionContext;
 use dp_block::StarknetVersion;
 use starknet_api::transaction::TransactionHash;
-use starknet_core::types::Felt;
 use starknet_core::types::TransactionTraceWithHash;
+use starknet_types_core::felt::Felt;
 use std::sync::Arc;
 
 // For now, we fallback to the sequencer - that is what pathfinder and juno do too, but this is temporary
@@ -30,7 +30,7 @@ pub async fn trace_transaction(
 
     let exec_context = ExecutionContext::new(Arc::clone(&starknet.backend), &block.info)?;
 
-    let mut block_txs = Iterator::zip(block.inner.transactions.iter(), block.info.tx_hashes())
+    let mut block_txs = Iterator::zip(block.inner.transactions.into_iter(), block.info.tx_hashes())
         .map(|(tx, hash)| to_blockifier_transactions(starknet, block.info.as_block_id(), tx, &TransactionHash(*hash)));
 
     // takes up until not including last tx