+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Next release
 
 - fix: fixed some readme stuff
+- test: add unitests primitives
+- tests: add tests for the rpcs endpoints
+- fix: pending contract storage not stored properly
+- test: add tests crate `db`
+- fix: --sync-l1-disabled cli option (#225)
 - feat: experimental block production and mempool
 - refactor: L1BlockMetric is intialized inside the EthereumClient new function
 - refactor: BlockMetrics divided in L1BlockMetrics and BlockMetrics